+++ AppleALC/kern_alc.cpp
@@ -2,187 +2,138 @@
 //  kern_alc.cpp
 //  AppleALC
 //
-//  Copyright © 2016 vit9696. All rights reserved.
+//  Copyright © 2016-2017 vit9696. All rights reserved.
 //
 
-#include "kern_alc.hpp"
-#include "kern_iokit.hpp"
-#include "kern_resources.hpp"
+#include <Headers/kern_api.hpp>
+#include <Library/LegacyIOService.h>
 
 #include <mach/vm_map.h>
 #include <IOKit/IORegistryEntry.h>
-#include <IOKit/IOService.h>
 
-//TODO: get rid of this
-static AlcEnabler *that {nullptr};
+#include "kern_alc.hpp"
+#include "kern_resources.hpp"
+
+// Only used in apple-driven callbacks
+static AlcEnabler *callbackAlc = nullptr;
+static KernelPatcher *callbackPatcher = nullptr;
 
 bool AlcEnabler::init() {
-	progressState = ProcessingState::NotReady;
-
-	patcher.init();
+	LiluAPI::Error error = lilu.onKextLoad(ADDPR(kextList), ADDPR(kextListSize),
+	[](void *user, KernelPatcher &patcher, size_t index, mach_vm_address_t address, size_t size) {
+		callbackAlc = static_cast<AlcEnabler *>(user);
+		callbackPatcher = &patcher;
+		callbackAlc->processKext(patcher, index, address, size);
+	}, this);
 	
-	if (patcher.getError() != KernelPatcher::Error::NoError) {
-		DBGLOG("alc @ failed to initialise kernel patcher");
-		patcher.clearError();
+	if (error != LiluAPI::Error::NoError) {
+		SYSLOG("alc @ failed to register onPatcherLoad method %d", error);
 		return false;
 	}
 	
-	return loadKexts();
+	return true;
 }
 
 void AlcEnabler::deinit() {
-	patcher.deinit();
 	controllers.deinit();
 	codecs.deinit();
 }
 
 void AlcEnabler::layoutLoadCallback(uint32_t requestTag, kern_return_t result, const void *resourceData, uint32_t resourceDataLength, void *context) {
-	if (that && that->orgLayoutLoadCallback) {
-		that->updateResource(Resource::Layout, result, resourceData, resourceDataLength);
-		that->orgLayoutLoadCallback(requestTag, result, resourceData, resourceDataLength, context);
+	if (callbackAlc && callbackPatcher && callbackAlc->orgLayoutLoadCallback) {
+		callbackAlc->updateResource(*callbackPatcher, Resource::Layout, result, resourceData, resourceDataLength);
+		callbackAlc->orgLayoutLoadCallback(requestTag, result, resourceData, resourceDataLength, context);
 	} else {
 		SYSLOG("alc @ layout callback arrived at nowhere");
 	}
 }
 
 void AlcEnabler::platformLoadCallback(uint32_t requestTag, kern_return_t result, const void *resourceData, uint32_t resourceDataLength, void *context) {
-	if (that && that->orgPlatformLoadCallback) {
-		that->updateResource(Resource::Platform, result, resourceData, resourceDataLength);
-		that->orgPlatformLoadCallback(requestTag, result, resourceData, resourceDataLength, context);
+	if (callbackAlc && callbackPatcher && callbackAlc->orgPlatformLoadCallback) {
+		callbackAlc->updateResource(*callbackPatcher, Resource::Platform, result, resourceData, resourceDataLength);
+		callbackAlc->orgPlatformLoadCallback(requestTag, result, resourceData, resourceDataLength, context);
 	} else {
 		SYSLOG("alc @ platform callback arrived at nowhere");
 	}
 }
 
-bool AlcEnabler::loadKexts() {
-	if (that) return true;
-	
-	for (size_t i = 0; i < kextListSize; i++) {
-		patcher.loadKinfo(&kextList[i]);
-		if (patcher.getError() != KernelPatcher::Error::NoError) {
-			SYSLOG("alc @ failed to load %s kext file", kextList[i].id);
-			patcher.clearError();
-			//return false;
-			// Depending on a system some kexts may actually not exist
-			continue;
-		}
-		
-		patcher.setupKextListening();
-		
-		if (patcher.getError() != KernelPatcher::Error::NoError) {
-			SYSLOG("alc @ failed to setup kext hooking");
-			patcher.clearError();
-			return false;
-		}
-		
-		auto handler = KernelPatcher::KextHandler::create(kextList[i].id, kextList[i].loadIndex,
-		[](KernelPatcher::KextHandler *h) {
-			if (h && that) {
-				that->processKext(h->index, h->address, h->size);
-			} else {
-				SYSLOG("alc @ notification callback arrived at nowhere");
+void AlcEnabler::processKext(KernelPatcher &patcher, size_t index, mach_vm_address_t address, size_t size) {
+	if (!(progressState & ProcessingState::ControllersLoaded)) {
+		grabControllers();
+		progressState |= ProcessingState::ControllersLoaded;
+	} else if (!(progressState & ProcessingState::CodecsLoaded)) {
+		for (size_t i = 0; i < ADDPR(kextListSize); i++) {
+			if (ADDPR(kextList)[i].loadIndex == index) {
+				if (ADDPR(kextList)[i].user[0] && grabCodecs()) {
+					progressState |= ProcessingState::CodecsLoaded;
+					break;
+				}
+				DBGLOG("alc @ failed to find a suitable codec, we have nothing to do");
+				// Continue to patch controllers
 			}
-		});
-		
-		if (!handler) {
-			SYSLOG("alc @ failed to allocate KextHandler");
-			return false;
 		}
-		
-		patcher.waitOnKext(handler);
-		
-		if (patcher.getError() != KernelPatcher::Error::NoError) {
-			SYSLOG("alc @ failed to wait on kext");
-			patcher.clearError();
-			KernelPatcher::KextHandler::deleter(handler);
-			return false;
+	}
+
+	if (progressState & ProcessingState::ControllersLoaded) {
+		for (size_t i = 0, num = controllers.size(); i < num; i++) {
+			auto &info = controllers[i]->info;
+			if (!info) {
+				DBGLOG("alc @ missing ControllerModInfo for %zu controller", i);
+				continue;
+			}
+			
+			applyPatches(patcher, index, info->patches, info->patchNum);
 		}
-	
 	}
 	
-	that = this;
-	return true;
-}
-
-void AlcEnabler::processKext(size_t index, mach_vm_address_t address, size_t size) {
-	patcher.updateRunningInfo(index, address, size);
-	
-	if (userPatchSize > 0) {
-		DBGLOG("alc @ applying user patches");
-		applyPatches(index, userPatch, userPatchSize);
+	if (progressState & ProcessingState::CodecsLoaded) {
+		for (size_t i = 0, num = codecs.size(); i < num; i++) {
+			auto &info = codecs[i]->info;
+			if (!info) {
+				SYSLOG("alc @ missing CodecModInfo for %zu codec", i);
+				continue;
+			}
+			
+			if (info->platforms > 0 && info->layoutNum > 0) {
+				DBGLOG("alc @ will route callbacks resource loading callbacks");
+				progressState |= ProcessingState::CallbacksWantRouting;
+			}
+			
+			applyPatches(patcher, index, info->patches, info->patchNum);
+		}
 	}
 	
-	if (patcher.getError() == KernelPatcher::Error::NoError) {
-		if (!(progressState & ProcessingState::ControllersLoaded)) {
-			grabControllers();
-			progressState |= ProcessingState::ControllersLoaded;
-		} else if (!(progressState & ProcessingState::CodecsLoaded)) {
-			for (size_t i = 0; i < kextListSize; i++) {
-				if (kextList[i].loadIndex == index) {
-					if (kextList[i].detectCodecs && grabCodecs()) {
-						progressState |= ProcessingState::CodecsLoaded;
-						break;
-					}
-					DBGLOG("alc @ failed to find a suitable codec, we have nothing to do");
-					// Continue to patch controllers
-				}
-			}
-		}
-	
-		if (progressState & ProcessingState::ControllersLoaded) {
-			for (size_t i = 0, num = controllers.size(); i < num; i++) {
-				auto &info = controllers[i]->info;
-				if (!info) {
-					DBGLOG("alc @ missing ControllerModInfo for %zu controller", i);
-					continue;
-				}
-				
-				applyPatches(index, info->patches, info->patchNum);
-			}
-		}
-		
-		if (progressState & ProcessingState::CodecsLoaded) {
-			for (size_t i = 0, num = codecs.size(); i < num; i++) {
-				auto &info = codecs[i]->info;
-				if (!info) {
-					SYSLOG("alc @ missing CodecModInfo for %zu codec", i);
-					continue;
-				}
-				
-				if (info->platforms > 0 && info->layoutNum > 0) {
-					DBGLOG("alc @ will route callbacks resource loading callbacks");
-					progressState |= ProcessingState::CallbacksWantRouting;
-				}
-				
-				applyPatches(index, info->patches, info->patchNum);
-			}
-		}
-		
-		if ((progressState & ProcessingState::CallbacksWantRouting) && !(progressState & ProcessingState::CallbacksRouted)) {
-			auto layout = patcher.solveSymbol(index, "__ZN14AppleHDADriver18layoutLoadCallbackEjiPKvjPv");
-			auto platform = patcher.solveSymbol(index, "__ZN14AppleHDADriver20platformLoadCallbackEjiPKvjPv");
+	if ((progressState & ProcessingState::CallbacksWantRouting) && !(progressState & ProcessingState::CallbacksRouted)) {
+		auto layout = patcher.solveSymbol(index, "__ZN14AppleHDADriver18layoutLoadCallbackEjiPKvjPv");
+		auto platform = patcher.solveSymbol(index, "__ZN14AppleHDADriver20platformLoadCallbackEjiPKvjPv");
 
-			if (!layout || !platform) {
-				SYSLOG("alc @ failed to find AppleHDA layout or platform callback symbols (%llX, %llX)", layout, platform);
-			} else if (orgLayoutLoadCallback = reinterpret_cast<t_callback>(patcher.routeFunction(layout, reinterpret_cast<mach_vm_address_t>(layoutLoadCallback), true)),
-					   patcher.getError() != KernelPatcher::Error::NoError) {
-				SYSLOG("alc @ failed to hook layout callback");
-			} else if (orgPlatformLoadCallback = reinterpret_cast<t_callback>(patcher.routeFunction(platform, reinterpret_cast<mach_vm_address_t>(platformLoadCallback), true)),
-					   patcher.getError() != KernelPatcher::Error::NoError) {
-				SYSLOG("alc @ failed to hook platform callback");
-			} else {
-				progressState |= ProcessingState::CallbacksRouted;
-			}
+		if (layout && platform) {
+			DBGLOG("layout call %X %X %X %X %X %X %X %X %X %X %X %X %X %X %X %X", ((uint8_t *)layout)[0], ((uint8_t *)layout)[1], ((uint8_t *)layout)[2], ((uint8_t *)layout)[3],
+				   ((uint8_t *)layout)[4], ((uint8_t *)layout)[5], ((uint8_t *)layout)[6], ((uint8_t *)layout)[7], ((uint8_t *)layout)[8], ((uint8_t *)layout)[9], ((uint8_t *)layout)[10],
+				   ((uint8_t *)layout)[11], ((uint8_t *)layout)[12], ((uint8_t *)layout)[13], ((uint8_t *)layout)[14], ((uint8_t *)layout)[15]);
+			DBGLOG("platform call %X %X %X %X %X %X %X %X %X %X %X %X %X %X %X %X", ((uint8_t *)platform)[0], ((uint8_t *)platform)[1], ((uint8_t *)platform)[2], ((uint8_t *)platform)[3],
+				   ((uint8_t *)platform)[4], ((uint8_t *)platform)[5], ((uint8_t *)platform)[6], ((uint8_t *)platform)[7], ((uint8_t *)platform)[8], ((uint8_t *)platform)[9], ((uint8_t *)platform)[10],
+				   ((uint8_t *)platform)[11], ((uint8_t *)platform)[12], ((uint8_t *)platform)[13], ((uint8_t *)platform)[14], ((uint8_t *)platform)[15]);
 		}
-	} else {
-		SYSLOG("alc @ failed to update kext running info");
+		
+		if (!layout || !platform) {
+			SYSLOG("alc @ failed to find AppleHDA layout or platform callback symbols (%llX, %llX)", layout, platform);
+		} else if (orgLayoutLoadCallback = reinterpret_cast<t_callback>(patcher.routeFunction(layout, reinterpret_cast<mach_vm_address_t>(layoutLoadCallback), true)),
+				   patcher.getError() != KernelPatcher::Error::NoError) {
+			SYSLOG("alc @ failed to hook layout callback");
+		} else if (orgPlatformLoadCallback = reinterpret_cast<t_callback>(patcher.routeFunction(platform, reinterpret_cast<mach_vm_address_t>(platformLoadCallback), true)),
+				   patcher.getError() != KernelPatcher::Error::NoError) {
+			SYSLOG("alc @ failed to hook platform callback");
+		} else {
+			progressState |= ProcessingState::CallbacksRouted;
+		}
 	}
 	
 	// Ignore all the errors for other processors
 	patcher.clearError();
 }
 
-void AlcEnabler::updateResource(Resource type, kern_return_t &result, const void * &resourceData, uint32_t &resourceDataLength) {
+void AlcEnabler::updateResource(KernelPatcher &patcher, Resource type, kern_return_t &result, const void * &resourceData, uint32_t &resourceDataLength) {
 	DBGLOG("alc @ resource-request arrived %s", type == Resource::Platform ? "paltform" : "layout");
 	
 	for (size_t i = 0, s = codecs.size(); i < s; i++) {
@@ -213,45 +164,40 @@ void AlcEnabler::updateResource(Resource type, kern_return_t &result, const void
 }
 
 void AlcEnabler::grabControllers() {
-	if (!that) {
-		SYSLOG("alc @ you should call grabCodecs right before AppleHDAController loading");
-		return;
-	}
-	
-	computerModel = IOUtil::getComputerModel();
+	computerModel = WIOKit::getComputerModel();
 	
 	bool found {false};
 	
-	for (size_t lookup = 0; lookup < codecLookupSize; lookup++) {
-		auto sect = IOUtil::findEntryByPrefix("/AppleACPIPlatformExpert", "PCI", gIOServicePlane);
+	for (size_t lookup = 0; lookup < ADDPR(codecLookupSize); lookup++) {
+		auto sect = WIOKit::findEntryByPrefix("/AppleACPIPlatformExpert", "PCI", gIOServicePlane);
 		
-		for (size_t i = 0; sect && i <= codecLookup[lookup].controllerNum; i++) {
-			sect = IOUtil::findEntryByPrefix(sect, codecLookup[lookup].tree[i], gIOServicePlane);
+		for (size_t i = 0; sect && i <= ADDPR(codecLookup)[lookup].controllerNum; i++) {
+			sect = WIOKit::findEntryByPrefix(sect, ADDPR(codecLookup)[lookup].tree[i], gIOServicePlane);
 			
-			if (sect && i == codecLookup[lookup].controllerNum) {
+			if (sect && i == ADDPR(codecLookup)[lookup].controllerNum) {
 				// Nice, we found some controller, add it
 				uint32_t ven {0}, dev {0}, rev {0}, platform {ControllerModInfo::PlatformAny}, lid {0};
 				
-				if (!IOUtil::getOSDataValue(sect, "vendor-id", ven) ||
-					!IOUtil::getOSDataValue(sect, "device-id", dev) ||
-					!IOUtil::getOSDataValue(sect, "revision-id", rev)) {
-					SYSLOG("alc @ found an incorrect controller at %s", codecLookup[lookup].tree[i]);
+				if (!WIOKit::getOSDataValue(sect, "vendor-id", ven) ||
+					!WIOKit::getOSDataValue(sect, "device-id", dev) ||
+					!WIOKit::getOSDataValue(sect, "revision-id", rev)) {
+					SYSLOG("alc @ found an incorrect controller at %s", ADDPR(codecLookup)[lookup].tree[i]);
 					break;
 				}
 				
-				if (codecLookup[lookup].detect && !IOUtil::getOSDataValue(sect, "layout-id", lid)) {
-					SYSLOG("alc @ layout-id was not provided by controller at %s", codecLookup[lookup].tree[i]);
+				if (ADDPR(codecLookup)[lookup].detect && !WIOKit::getOSDataValue(sect, "layout-id", lid)) {
+					SYSLOG("alc @ layout-id was not provided by controller at %s", ADDPR(codecLookup)[lookup].tree[i]);
 					break;
 				}
 				
-				if (IOUtil::getOSDataValue(sect, "AAPL,ig-platform-id", platform)) {
-					DBGLOG("alc @ AAPL,ig-platform-id %X was found in controller at %s", platform, codecLookup[lookup].tree[i]);
+				if (WIOKit::getOSDataValue(sect, "AAPL,ig-platform-id", platform)) {
+					DBGLOG("alc @ AAPL,ig-platform-id %X was found in controller at %s", platform, ADDPR(codecLookup)[lookup].tree[i]);
 				}
 				
-				auto controller = ControllerInfo::create(ven, dev, rev, platform, lid, codecLookup[lookup].detect);
+				auto controller = ControllerInfo::create(ven, dev, rev, platform, lid, ADDPR(codecLookup)[lookup].detect);
 				if (controller) {
 					if (controllers.push_back(controller)) {
-						controller->lookup = &codecLookup[lookup];
+						controller->lookup = &ADDPR(codecLookup)[lookup];
 						found = true;
 					} else {
 						SYSLOG("alc @ failed to store controller info for %X:%X:%X", ven, dev, rev);
@@ -271,11 +217,6 @@ void AlcEnabler::grabControllers() {
 }
 
 bool AlcEnabler::grabCodecs() {
-	if (!that) {
-		SYSLOG("alc @ you should call grabCodecs right before AppleHDA loading");
-		return false;
-	}
-	
 	for (currentController = 0; currentController < controllers.size(); currentController++) {
 		auto ctlr = controllers[currentController];
 		
@@ -283,13 +224,15 @@ bool AlcEnabler::grabCodecs() {
 		if (!ctlr->detect)
 			continue;
 
-		auto sect = IOUtil::findEntryByPrefix("/AppleACPIPlatformExpert", "PCI", gIOServicePlane);
+		auto sect = WIOKit::findEntryByPrefix("/AppleACPIPlatformExpert", "PCI", gIOServicePlane);
 
 		for (size_t i = 0; sect && i < ctlr->lookup->treeSize; i++) {
 			bool last = i+1 == ctlr->lookup->treeSize;
-			sect = IOUtil::findEntryByPrefix(sect, ctlr->lookup->tree[i], gIOServicePlane,
-											 last ? [](IORegistryEntry *e) {
-				
+			sect = WIOKit::findEntryByPrefix(sect, ctlr->lookup->tree[i], gIOServicePlane,
+											 last ? [](void *user, IORegistryEntry *e) {
+	
+				auto alc = static_cast<AlcEnabler *>(user);
+												 
 				auto ven = e->getProperty("IOHDACodecVendorID");
 				auto rev = e->getProperty("IOHDACodecRevisionID");
 				
@@ -307,11 +250,11 @@ bool AlcEnabler::grabCodecs() {
 					return true;
 				}
 				
-				auto ci = AlcEnabler::CodecInfo::create(that->currentController,
+				auto ci = AlcEnabler::CodecInfo::create(alc->currentController,
 														venNum->unsigned64BitValue(),
 														revNum->unsigned32BitValue());
 				if (ci) {
-					if (!that->codecs.push_back(ci)) {
+					if (!alc->codecs.push_back(ci)) {
 						SYSLOG("alc @ failed to store codec info for %X:%X:%X", ci->vendor, ci->codec, ci->revision);
 						AlcEnabler::CodecInfo::deleter(ci);
 					}
@@ -321,7 +264,7 @@ bool AlcEnabler::grabCodecs() {
 				
 				return true;
 			
-			} : nullptr, last);
+			} : nullptr, last, this);
 		}
 	}
 
@@ -331,34 +274,34 @@ bool AlcEnabler::grabCodecs() {
 void AlcEnabler::validateControllers() {
 	for (size_t i = 0, num = controllers.size(); i < num; i++) {
 		DBGLOG("alc @ validating %zu controller %X:%X:%X", i, controllers[i]->vendor, controllers[i]->device, controllers[i]->revision);
-		for (size_t mod = 0; mod < controllerModSize; mod++) {
-			DBGLOG("alc @ comparing to %zu mod %X:%X", mod, controllerMod[mod].vendor, controllerMod[mod].device);
-			if (controllers[i]->vendor == controllerMod[mod].vendor &&
-				controllers[i]->device == controllerMod[mod].device) {
+		for (size_t mod = 0; mod < ADDPR(controllerModSize); mod++) {
+			DBGLOG("alc @ comparing to %zu mod %X:%X", mod, ADDPR(controllerMod)[mod].vendor, ADDPR(controllerMod)[mod].device);
+			if (controllers[i]->vendor == ADDPR(controllerMod)[mod].vendor &&
+				controllers[i]->device == ADDPR(controllerMod)[mod].device) {
 				
 				// Check revision if present
 				size_t rev {0};
-				while (rev < controllerMod[mod].revisionNum &&
-					   controllerMod[mod].revisions[rev] != controllers[i]->revision)
+				while (rev < ADDPR(controllerMod)[mod].revisionNum &&
+					   ADDPR(controllerMod)[mod].revisions[rev] != controllers[i]->revision)
 					rev++;
 				
 				// Check AAPL,ig-platform-id if present
-				if (controllerMod[mod].platform != ControllerModInfo::PlatformAny &&
-					controllerMod[mod].platform != controllers[i]->platform) {
-					DBGLOG("alc @ not matching platform was found %X vs %X", controllerMod[mod].platform, controllers[i]->platform);
+				if (ADDPR(controllerMod)[mod].platform != ControllerModInfo::PlatformAny &&
+					ADDPR(controllerMod)[mod].platform != controllers[i]->platform) {
+					DBGLOG("alc @ not matching platform was found %X vs %X", ADDPR(controllerMod)[mod].platform, controllers[i]->platform);
 					continue;
 				}
 				
 				// Check if computer model is suitable
-				if (!(computerModel & controllerMod[mod].computerModel)) {
-					DBGLOG("alc @ unsuitable computer model was found %X vs %X", controllerMod[mod].computerModel, computerModel);
+				if (!(computerModel & ADDPR(controllerMod)[mod].computerModel)) {
+					DBGLOG("alc @ unsuitable computer model was found %X vs %X", ADDPR(controllerMod)[mod].computerModel, computerModel);
 					continue;
 				}
 			
-				if (rev != controllerMod[mod].revisionNum ||
-					controllerMod[mod].revisionNum == 0) {
+				if (rev != ADDPR(controllerMod)[mod].revisionNum ||
+					ADDPR(controllerMod)[mod].revisionNum == 0) {
 					DBGLOG("alc @ found mod for %zu controller", i);
-					controllers[i]->info = &controllerMod[mod];
+					controllers[i]->info = &ADDPR(controllerMod)[mod];
 					break;
 				}
 			}
@@ -374,35 +317,35 @@ bool AlcEnabler::validateCodecs() {
 		
 		// Check vendor
 		size_t vIdx {0};
-		while (vIdx < vendorModSize && vendorMod[vIdx].vendor != codecs[i]->vendor)
+		while (vIdx < ADDPR(vendorModSize) && ADDPR(vendorMod)[vIdx].vendor != codecs[i]->vendor)
 			vIdx++;
 		
-		if (vIdx != vendorModSize) {
+		if (vIdx != ADDPR(vendorModSize)) {
 			// Check codec
 			size_t cIdx {0};
-			while (cIdx < vendorMod[vIdx].codecsNum &&
-				   vendorMod[vIdx].codecs[cIdx].codec != codecs[i]->codec)
+			while (cIdx < ADDPR(vendorMod)[vIdx].codecsNum &&
+				   ADDPR(vendorMod)[vIdx].codecs[cIdx].codec != codecs[i]->codec)
 				cIdx++;
 			
-			if (cIdx != vendorMod[vIdx].codecsNum) {
+			if (cIdx != ADDPR(vendorMod)[vIdx].codecsNum) {
 				// Check revision if present
 				size_t rIdx {0};
-				while (rIdx < vendorMod[vIdx].codecs[cIdx].revisionNum &&
-					   vendorMod[vIdx].codecs[cIdx].revisions[rIdx] != codecs[i]->revision)
+				while (rIdx < ADDPR(vendorMod)[vIdx].codecs[cIdx].revisionNum &&
+					   ADDPR(vendorMod)[vIdx].codecs[cIdx].revisions[rIdx] != codecs[i]->revision)
 					rIdx++;
 				
-				if (rIdx != vendorMod[vIdx].codecs[cIdx].revisionNum ||
-					vendorMod[vIdx].codecs[cIdx].revisionNum == 0) {
-					codecs[i]->info = &vendorMod[vIdx].codecs[cIdx];
+				if (rIdx != ADDPR(vendorMod)[vIdx].codecs[cIdx].revisionNum ||
+					ADDPR(vendorMod)[vIdx].codecs[cIdx].revisionNum == 0) {
+					codecs[i]->info = &ADDPR(vendorMod)[vIdx].codecs[cIdx];
 					suitable = true;
 					
 				}
 				
 				DBGLOG("alc @ found %s %s %s codec revision 0x%X",
-					   suitable ? "supported" : "unsupported", vendorMod[vIdx].name,
-					   vendorMod[vIdx].codecs[cIdx].name, codecs[i]->revision);
+					   suitable ? "supported" : "unsupported", ADDPR(vendorMod)[vIdx].name,
+					   ADDPR(vendorMod)[vIdx].codecs[cIdx].name, codecs[i]->revision);
 			} else {
-				DBGLOG("alc @ found unsupported %s codec 0x%X revision 0x%X", vendorMod[vIdx].name,
+				DBGLOG("alc @ found unsupported %s codec 0x%X revision 0x%X", ADDPR(vendorMod)[vIdx].name,
 					   codecs[i]->codec, codecs[i]->revision);
 			}
 		} else {
@@ -418,7 +361,7 @@ bool AlcEnabler::validateCodecs() {
 	return codecs.size() > 0;
 }
 
-void AlcEnabler::applyPatches(size_t index, const KextPatch *patches, size_t patchNum) {
+void AlcEnabler::applyPatches(KernelPatcher &patcher, size_t index, const KextPatch *patches, size_t patchNum) {
 	DBGLOG("alc @ applying patches for %zu kext", index);
 	for (size_t p = 0; p < patchNum; p++) {
 		auto &patch = patches[p];